Publish the measured rungs, and keep all three
The study ran against rendered prose as well as drawn fixtures, and the
numbers it was written to settle survived it. 1024, 1568 and 2576 stay.

Stem retention on real letterforms, at a viewport wide enough that all three
rungs genuinely shrink: `max` returns everything at every font size tested,
`detail` recovers ordinary body copy, and `default` keeps headings and layout
intact while damaging small body copy. That last one is the lever working
rather than a defect — a caller that needs to read something is pushed toward
the snapshot or the evaluation, which return text and cost almost nothing.

The property underneath is now measured rather than asserted, and it has a
mechanism: a downscale destroys a feature once the feature's period falls below
roughly two and a half destination pixels. So what a rung costs depends on the
size of the feature rather than on the rung, which is why a block-scale feature
survives the whole ladder and fine text detail does not.

What the study did NOT settle is written down beside what it did. The
instruments measure what the pipeline destroys, which bounds what a reader
could recover but does not predict what one will; the absolute legibility floor
is still open and §9.3 keeps it open. A rendered-text suite carries the real
letterforms and skips per-test where there is no browser, so a green pipeline
is not read as evidence it ran.

Two tests are added from mutation survivors rather than from a checklist. The
ratio is now proved on a portrait picture, because on a landscape one "the cap
over the source width" and "the written width over the source width" agree —
so every existing fixture would have passed a ratio computed from the cap, and
a tall page is exactly the shape a full-page capture makes. The other pins
which of several tying rows is reported.

+++ tests/helpers/ladder-fixtures.ts
@@ -53,7 +53,7 @@ export function paint(
   width: number,
   height: number,
 ): void {
-  const data = image.pixels as Uint8Array;
+  const data = image.pixels;
   for (let row = 0; row < height; row++) {
     const atY = y + row;
     if (atY < 0 || atY >= image.height) continue;
